super1

+++ database/migrations/2024_01_28_000003_create_users_table.php
@@ -8,35 +8,34 @@ return new class extends Migration
 {
     public function up()
     {
-        Schema::create('users', function (Blueprint $table) {
-            $table->id();
-            $table->string('first_name');
-            $table->string('last_name');
-            $table->string('utis_code')->unique();
-            $table->string('email')->unique();
-            $table->string('username')->unique();
-            $table->string('password');
-            $table->enum('user_type', ['superadmin', 'sectoradmin', 'schooladmin']);
-            $table->foreignId('region_id')->nullable()->constrained()->onDelete('cascade');
-            $table->foreignId('sector_id')->nullable()->constrained()->onDelete('cascade');
-            $table->foreignId('school_id')->nullable()->constrained()->onDelete('cascade');
-            
-            // New login tracking columns
-            $table->timestamp('last_login_at')->nullable();
-            $table->string('last_login_ip')->nullable();
-            
-            $table->timestamp('email_verified_at')->nullable();
-            $table->rememberToken();
-            $table->timestamps();
-            $table->softDeletes();
-        });
+        if (!Schema::hasTable('users')) {
+            Schema::create('users', function (Blueprint $table) {
+                $table->id();
+                $table->string('first_name');
+                $table->string('last_name');
+                $table->string('utis_code')->unique();
+                $table->string('email')->unique();
+                $table->string('username')->unique();
+                $table->string('password');
+                $table->enum('user_type', ['superadmin', 'sector-admin', 'school-admin']);
+                $table->foreignId('region_id')->nullable()->constrained()->onDelete('cascade');
+                $table->foreignId('sector_id')->nullable()->constrained()->onDelete('cascade');
+                $table->foreignId('school_id')->nullable()->constrained()->onDelete('cascade');
+                
+                // New login tracking columns
+                $table->timestamp('last_login_at')->nullable();
+                $table->string('last_login_ip')->nullable();
+                
+                $table->timestamp('email_verified_at')->nullable();
+                $table->rememberToken();
+                $table->timestamps();
+                $table->softDeletes();
+            });
+        }
     }
 
     public function down()
     {
-        Schema::table('users', function (Blueprint $table) {
-            // If you want to drop the newly added columns during rollback
-            $table->dropColumn(['last_login_at', 'last_login_ip']);
-        });
+        Schema::dropIfExists('users');
     }
 };
